+++ packages/core/src/actions/enqueueActions.ts
@@ -12,6 +12,7 @@ import {
   MachineContext,
   ParameterizedObject,
   ProvidedActor,
+  BuiltinActionResolution,
   UnifiedArg
 } from '../types.ts';
 import { assign } from './assign.ts';
@@ -130,7 +131,7 @@ function resolveEnqueueActions(
       EventObject
     >;
   }
-) {
+): BuiltinActionResolution {
   const actions: any[] = [];
   const enqueue: Parameters<typeof collect>[0]['enqueue'] = function enqueue(
     action
@@ -311,8 +312,8 @@ export function enqueueActions<
   TEmitted
 > {
   function enqueueActions(
-    args: ActionArgs<TContext, TExpressionEvent, TEvent>,
-    params: unknown
+    _args: ActionArgs<TContext, TExpressionEvent, TEvent>,
+    _params: unknown
   ) {
     if (isDevelopment) {
       throw new Error(`This isn't supposed to be called`);