+++ packages/xstate-store/test/store.test.ts
@@ -1,5 +1,6 @@
 import { produce } from 'immer';
 import { createStore, createStoreWithProducer } from '../src/index.ts';
+import { createBrowserInspector } from '@statelyai/inspect';
 
 it('updates a store with an event without mutating original context', () => {
   const context = { count: 0 };
@@ -120,6 +121,28 @@ it('createStoreWithProducer(…) works with an immer producer', () => {
   expect(store.getInitialSnapshot().context).toEqual({ count: 0 });
 });
 
+it('createStoreWithProducer(…) works with an immer producer (object API)', () => {
+  const store = createStoreWithProducer(produce, {
+    context: {
+      count: 0
+    },
+    on: {
+      inc: (ctx, ev: { by: number }) => {
+        ctx.count += ev.by;
+      }
+    }
+  });
+
+  store.send({ type: 'inc', by: 3 });
+  store.send({
+    // @ts-expect-error
+    type: 'whatever'
+  });
+
+  expect(store.getSnapshot().context).toEqual({ count: 3 });
+  expect(store.getInitialSnapshot().context).toEqual({ count: 0 });
+});
+
 it('createStoreWithProducer(…) infers the context type properly with a producer', () => {
   const store = createStoreWithProducer(
     produce,
@@ -132,6 +155,21 @@ it('createStoreWithProducer(…) infers the context type properly with a produce
       }
     }
   );
+
+  store.getSnapshot().context satisfies { count: number };
+});
+
+it('createStoreWithProducer(…) infers the context type properly with a producer (object API)', () => {
+  const store = createStoreWithProducer(produce, {
+    context: {
+      count: 0
+    },
+    on: {
+      inc: (ctx, ev: { by: number }, enq) => {
+        ctx.count += ev.by;
+      }
+    }
+  });
 
   store.getSnapshot().context satisfies { count: number };
 });
@@ -202,6 +240,19 @@ it('can be inspected', () => {
       snapshot: expect.objectContaining({ context: { count: 1 } })
     })
   ]);
+});
+
+it('inspection with @statelyai/inspect typechecks correctly', () => {
+  const store = createStore({
+    context: {},
+    on: {}
+  });
+
+  const inspector = createBrowserInspector({
+    autoStart: false
+  });
+
+  store.inspect(inspector.inspect);
 });
 
 it('emitted events can be subscribed to', () => {